+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
  - Logging
+ - Read single job endpoint
+ - Read all jobs endpoint
 
 ## [0.1.1]
 ### Fixed